+++ package.json
@@ -2,7 +2,7 @@
   "name": "alexa2mqtt",
   "version": "0.0.1",
   "description": "Nodejs port of github.com/armzilla/amazon-echo-ha-bridge",
-  "main": "index.js",
+  "main": "alexa2mqtt.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
